requirements.txt
torch>=2.2
transformers>=4.45
peft>=0.12
datasets>=2.20
PyYAML>=6.0
tqdm>=4.66
accelerate>=0.33
safetensors>=0.4
sentencepiece>=0.2
numpy>=1.26
sympy>=1.12
deepspeed>=0.15
wandb>=0.16
geoopt>=0.5
